Reorganize  tests
